build: run publish sanity checks before building (#13060)
Runs the publish sanity checks before building the release output of Angular Material. Building the release output is very time consuming, and therefore the sanity checks should run before.

Closes #12918
