+++ tools/gulp/gulpfile.ts
@@ -33,4 +33,5 @@ import './tasks/unit-test';
 import './tasks/universal';
 
 import './tasks/publish/publish-task';
+import './tasks/publish/sanity-checks';
 import './tasks/publish/validate-release';
