+++ tools/gulp/tasks/publish/publish-task.ts
@@ -1,4 +1,4 @@
-import {green, grey, red, yellow} from 'chalk';
+import {green, grey, yellow} from 'chalk';
 import {spawn} from 'child_process';
 import {existsSync, statSync} from 'fs-extra';
 import {task} from 'gulp';
@@ -6,7 +6,6 @@ import {buildConfig, sequenceTask} from 'material2-build-tools';
 import * as minimist from 'minimist';
 import {join} from 'path';
 import {execTask} from '../../util/task_helpers';
-import {checkPublishBranch, versionNameRegex} from './branch-check';
 
 /** Packages that will be published to NPM by the release task. */
 export const releasePackages = [
@@ -21,9 +20,9 @@ export const releasePackages = [
 const argv = minimist(process.argv.slice(3));
 
 task('publish', sequenceTask(
+  ':publish:sanity-checks',
   ':publish:whoami',
   ':publish:build-releases',
-  'validate-release:check-remote-tag',
   'validate-release:check-bundles',
   ':publish',
   ':publish:logout',
@@ -48,14 +47,6 @@ task(':publish', async () => {
   const version = buildConfig.projectVersion;
   const currentDir = process.cwd();
 
-  if (!version.match(versionNameRegex)) {
-    console.error(red(`Error: Cannot publish due to an invalid version name. Version ` +
-        `"${version}" is not following our semver format.`));
-    console.error(yellow(`A version should follow this format: X.X.X, X.X.X-beta.X, ` +
-        `X.X.X-alpha.X, X.X.X-rc.X`));
-    return;
-  }
-
   console.log();
   if (!tag) {
     console.log(grey('> You can specify the tag by passing --tag=labelName.\n'));
@@ -65,20 +56,11 @@ task(':publish', async () => {
   }
   console.log();
 
-  if (version.match(/(alpha|beta|rc)/) && (!tag || tag === 'latest')) {
-    console.error(red(`Publishing ${version} to the "latest" tag is not allowed.`));
-    console.error(red(`Alpha, Beta or RC versions shouldn't be published to "latest".`));
-    console.error();
-    return;
-  }
-
   if (releasePackages.length > 1) {
     console.warn(yellow('Warning: Multiple packages will be released.'));
     console.warn(yellow('Warning: Packages to be released:', releasePackages.join(', ')));
     console.warn();
   }
-
-  checkPublishBranch(version);
 
   console.log(yellow('> Make sure to check the "requiredAngularVersion" in the package.json.'));
   console.log(yellow('> The version in the config defines the peer dependency of Angular.'));
